requirements.txt
requests
requests-cache
beautifulsoup4
tenacity
fastapi
pydantic-settings
beanie
uvicorn[standard]